refactor(cli): shared BAKIN_URL-aware http client; fix schedule BAKIN_URL bug
Extracts the CLI HTTP client to src/cli/http.ts — one BASE_URL resolution
(process.env.BAKIN_URL || localhost:PORT) plus api/apiGet/apiPost/apiPostJson/apiDelete,
the jsonObject/parseJsonText/apiErrorPayload/isServerConnectionError guards, and
getCliAgent/getCliRoster — moved verbatim out of cli/bakin.ts (which now imports them;
call sites unchanged). Only fetch + the lightweight api-error helpers are imported, so
nothing new enters the binary entry's import graph.

Fixes the real bug: src/cli/schedule.ts hardcoded "http://localhost:PORT" and ignored
BAKIN_URL, so "BAKIN_URL=... bakin schedule list" silently hit the wrong host. It now
imports the shared BASE_URL. Its thin schedule-prefixed typed wrappers stay (they use a
distinct 'API error' prefix; folding them into the generic client would change error text)
— the duplicated, buggy BASE_URL is gone.

Deferred (noted): replacing isServerConnectionError's message-text classification with a
typed connection error — moved verbatim for now; the typed-error improvement is a separate,
optional follow-up.

Verified: typecheck + lint clean; tests/cli 252/0 (--isolate); full binary build.
